requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
mysql-connector-python==9.0.0
python-dotenv==1.0.1
